Merged my terrible DB project in.
Note : There is a lot of stuff from it that needs to be deleted and there needs to be a more recently version of the DB handler.
I am working on making the login work again on Postgres so you guys can try it.
I might deploy this on my box so you don't need to setup posgres yourselfs.

sorry for the wait. pushing from my chromebook